Replace fkyaml with eckit, fix CI build system, and upgrade Node.js
- Removed vendored fkyaml library.
- Rewrote C++ YAML bridge (yaml_reader.cpp) using eckit APIs with caching and flexible type handling to ensure robustness and performance.
- Updated root and component CMakeLists.txt to find and link against eckit.
- Updated CI workflows to use NOAA-EMC/spack fork with required normalization script.
- Upgraded CI workflows to Node.js 24 and fixed Intel GPG key handling.
- Switched CI Doxygen installation to apt in both setup and build jobs to avoid 'Illegal instruction' errors.
- Ensured Intel environment variables are sourced before Spack configuration in the Intel workflow.
- Updated ci/spack.yaml to include eckit dependency and allow flexible concretization.
- Added eckit to modulefiles/upp_common.lua.
- Modernized GitHub Actions cache keys by hashing Spack configuration, modulefiles, and workflow files.
- Fixed vector type mismatch in yaml_reader.cpp (eckit uses int for getIntVector).
- Verified local compilation of the C++ bridge.

+++ sorc/ncep_post.fd/yaml_reader.cpp
@@ -168,7 +168,7 @@ int yaml_get_param_array_size(int pset_idx, int param_idx, const char* key) {
                 return (int)v.size();
             } catch (...) {
                 try {
-                    std::vector<long> v = cached_params[param_idx].getIntVector(key);
+                    std::vector<int> v = cached_params[param_idx].getIntVector(key);
                     return (int)v.size();
                 } catch (...) {
                     try {
@@ -201,7 +201,7 @@ void yaml_get_param_array_float(int pset_idx, int param_idx, const char* key, fl
                 for (int i = 0; i < out_len && i < (int)v.size(); ++i) out[i] = (float)v[i];
             } catch (...) {
                 try {
-                    std::vector<long> v = param.getIntVector(key);
+                    std::vector<int> v = param.getIntVector(key);
                     for (int i = 0; i < out_len && i < (int)v.size(); ++i) out[i] = (float)v[i];
                 } catch (...) {
                     try {
@@ -232,7 +232,7 @@ void yaml_get_param_array_int(int pset_idx, int param_idx, const char* key, int*
         const auto& param = cached_params[param_idx];
         if (param.has(key)) {
             try {
-                std::vector<long> v = param.getIntVector(key);
+                std::vector<int> v = param.getIntVector(key);
                 for (int i = 0; i < out_len && i < (int)v.size(); ++i) out[i] = (int)v[i];
             } catch (...) {
                 try {
